Added types to metamui-primitives and removed collective's unused imports
