+++ metamui-primitives/src/traits.rs
@@ -4,9 +4,7 @@ use sp_core::hexdisplay::HexDisplay;
 use sp_runtime::RuntimeDebug;
 use scale_info::TypeInfo;
 use frame_support::{traits::{ConstU32}, BoundedVec, sp_runtime::DispatchError};
-use sp_core::sr25519::{Signature as SRSignature};
 use sp_std::{prelude::*};
-
 
 // DID
 
@@ -78,51 +76,7 @@ impl<AccountId: Default> Default for MultiAddress<AccountId> {
         MultiAddress::Id(Default::default())
     }
 }
-
-
 // VC
-
-/// VC Property max length
-pub type VCPropertyLimit = ConstU32<32>;
-/// VC Property type
-pub type VCProperty = BoundedVec<u8, VCPropertyLimit>;
-
-/// Type of VCs
-#[derive(Encode, Decode, Clone, PartialEq, Eq, Debug, TypeInfo)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-pub enum VCType {
-    /// VC to create a Token
-    TokenVC,
-    /// VC to slash token
-    SlashTokens,
-    /// VC to mint token
-    MintTokens,
-    /// VC to transfer token
-    TokenTransferVC,
-    /// VC for generic purpose
-    GenericVC,
-}
-
-
-/// Struct for VC
-#[derive(Clone, PartialEq, Eq, RuntimeDebug, Encode, Decode, TypeInfo)]
-pub struct VC<Hash> {
-    /// Hash of the data in VC
-    pub hash: Hash,
-    /// Owner of VC
-    pub owner: Did,
-    /// Issuers of VC
-    pub issuers: Vec<Did>,
-    /// Signatures of Issuers on hash
-    pub signatures: Vec<SRSignature>,
-    /// If VC is used or not
-    pub is_vc_used: bool,
-    /// Type of VC
-    pub vc_type: VCType,
-    /// VC payload
-    pub vc_property: VCProperty,
-}
-
 /// Trait to get VC details
 pub trait VCResolve<Hash> {
     /// Get VC from VC Id
@@ -133,9 +87,8 @@ pub trait VCResolve<Hash> {
     fn set_vc_used(vc_id: &VCid, is_vc_used: bool);
     /// Decode VC
     fn decode_vc<E: Decode>(vc_bytes: &[u8]) -> Result<E, DispatchError>;
-}
-
-
+  }
+  
 impl<Hash> VCResolve<Hash> for () {
     /// Get VC from VC Id
     fn get_vc(_vc_id: &VCid) -> Option<VC<Hash>> {
@@ -152,6 +105,22 @@ impl<Hash> VCResolve<Hash> for () {
     /// Decode VC
     fn decode_vc<E: Decode>(_vc_bytes: &[u8]) -> Result<E, DispatchError> {
         Err("Not Implemented".into())
+    }
+}
+
+pub trait HasVCId {
+    fn vc_id(&self) -> VCid;
+}
+
+impl HasVCId for SlashMintTokens {
+    fn vc_id(&self) -> VCid {
+        self.vc_id
+    }
+}
+
+impl HasVCId for TokenTransferVC {
+    fn vc_id(&self) -> VCid {
+        self.vc_id
     }
 }
